hw/hw.week 6 (중간)/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
pydantic-settings==2.5.2
pydantic[email]==2.9.2
python-multipart==0.0.12
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.51.2
sentry-sdk[fastapi]==2.16.0
slowapi==0.1.9
python-dotenv==1.0.1
cryptography==43.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
anyio[trio]
